Add commands to extract members, add members in another channel, group, etc.
Add new operations to the TelePilot node and trigger node.

* **common.descriptions.ts**
  - Add `promoteGroupMember`, `demoteGroupMember`, `muteGroupMember`, `unmuteGroupMember`, and `createGroupPoll` operations to the `operationGroup` options.
  - Add `sendLocationMessage`, `sendContactInformation`, `createCustomKeyboardButtons`, `sendVideoMessageWithCaption`, `sendSticker`, and `sendGif` operations to the `operationChat` options.

* **TelePilot.node.ts**
  - Implement `addChatMembers`, `promoteGroupMember`, `demoteGroupMember`, `muteGroupMember`, `unmuteGroupMember`, `createGroupPoll`, `sendLocationMessage`, `sendContactInformation`, `createCustomKeyboardButtons`, `sendVideoMessageWithCaption`, `sendSticker`, `sendGif`, `extractChannelMembers`, `extractGroupMembers`, and `forceEntryChannel` operations.

* **TelePilotTrigger.node.ts**
  - Implement `addChatMembers`, `promoteGroupMember`, `demoteGroupMember`, `muteGroupMember`, `unmuteGroupMember`, `createGroupPoll`, `sendLocationMessage`, `sendContactInformation`, `createCustomKeyboardButtons`, `sendVideoMessageWithCaption`, `sendSticker`, `sendGif`, `extractChannelMembers`, `extractGroupMembers`, and `forceEntryChannel` operations.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/crialabs/n8n-nodes-telepilot?shareId=XXXX-XXXX-XXXX-XXXX).
